Changes made:
Created two methods in the Controller class which creates a mineral arraylist of the names of all the minerals. Other stylistic code changes were made in other classes. Created a TopTen java class that is for testing purposes only. Made some changes to DatabaseHandler which are important for SQLite integration.